more gold

+++ sator-reward/program/src/types.rs
@@ -10,6 +10,6 @@ use solana_program::pubkey::Pubkey;
 #[derive(Debug, BorshDeserialize, BorshSerialize, BorshSchema, Default, Clone, Copy)]
 pub struct Winner {
     pub user_wallet: Pubkey,
-    pub points: u16,
+    pub points: u32,
     pub claimed: bool,
 }
